+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Improvements
 + Replaced dependency `find-up` with `findup-sync` in `algob`.
 + Added `algopy` in `@algo-builder/algob/sample-project`, which enables users to pass template parameters to PyTEAL contracts. Updated docs.
++ TEALv3 support in `@algo-builder/runtime`: `assert`, `pushint`, `pushbytes`, `swap` opcodes.
 + Update runtime to process execParams.deployASA, deploySSC, OptInToASA, OptIntoSSC
 
 ### Bug fixes